Disabled interrupts during context save.
There's some sort of nondeterminism if I don't do this.  I can't
explain it, mostly because I haven't finished thinking about it,
but if I leave interrupts enabled the system hangs within about
a minute.

A hypothesis: one of the task stacks may be too shallow to support
taking an interrupt while the full context is saved.