Fixed a precedence bug in Task initialization.
Take line 127 of src/task.cc for example:
  _PUSH(code >> 8);

'code' is a uintptr_t, so we can reasonably shift it and expect to
get high bits.  However, this is a macro, and it expands to
  *(sp--) = (uint8_t) code >> 8;

The compiler sees that the result can only be zero and hardcodes it.
As a result, any task whose main loop started higher than 0x1FF would
fail to initialize.